use expiry from uri in wcpairing
