+++ Sources/WalletConnectPairing/Services/App/AppPairService.swift
@@ -15,8 +15,9 @@ actor AppPairService {
         let topic = String.generateTopic()
         try await networkingInteractor.subscribe(topic: topic)
         let symKey = try! kms.createSymmetricKey(topic)
-        let pairing = WCPairing(topic: topic)
-        let uri = WalletConnectURI(topic: topic, symKey: symKey.hexRepresentation, relay: pairing.relay)
+        let relay = RelayProtocolOptions(protocol: "irn", data: nil)
+        let uri = WalletConnectURI(topic: topic, symKey: symKey.hexRepresentation, relay: relay)
+        let pairing = WCPairing(uri: uri)
         pairingStorage.setPairing(pairing)
         return uri
     }
